uuid and mechanisms

+++ chart_of_accounts/views.py
@@ -4,15 +4,24 @@ from .forms import UploadFileForm
 from .models import Chart_of_accounts
 from .forms import AccountForm
 from django.http import JsonResponse
+from django.contrib import messages
 
 def upload_and_save(request):
     if request.method == 'POST':
         form = UploadFileForm(request.POST, request.FILES)
         if form.is_valid():
-            Chart_of_accounts.objects.all().delete()  # Apaga todos os registros existentes
             excel_file = request.FILES['file']
-            wb = openpyxl.load_workbook(excel_file)
+            wb = openpyxl.load_workbook(excel_file, data_only=True)
             worksheet = wb.active
+
+            # Verificar as células especiais
+            if (worksheet['XFD1'].value != "sppiff" or
+                worksheet['XFD2'].value != "1ccd101b-aed4-4b56-b852-02c8ce8a6f70" or
+                worksheet['XFD3'].value != "a5296f55-e314-48d2-8415-f0204d3162f8"):
+                messages.error(request, 'A planilha não passou na verificação de segurança.')
+                return redirect('/plano_de_contas/')  # Substitua pelo nome da URL de upload
+
+            Chart_of_accounts.objects.all().delete()  # Apaga todos os registros existentes
             
             # Criar um objeto para cada linha na planilha
             for row in worksheet.iter_rows(min_row=2, values_only=True):
@@ -44,6 +53,18 @@ def upload_and_save(request):
     
     chart_of_accounts = Chart_of_accounts.objects.all().order_by('-group')
     return render(request, 'chart_of_accounts.html', {'form': form, 'chart_of_accounts': chart_of_accounts})
+
+def verify_special_cells(worksheet):
+    expected_values = {
+        'XFD1': "sppiff",
+        'XFD2': "1ccd101b-aed4-4b56-b852-02c8ce8a6f70",
+        'XFD3': "a5296f55-e314-48d2-8415-f0204d3162f8"
+    }
+    for cell, expected_value in expected_values.items():
+        actual_value = worksheet[cell].value
+        if actual_value != expected_value:
+            return False
+    return True
 
 def add_account(request):
     if request.method == 'POST':
